third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/open-iframe-src-01-expected.txt

CONSOLE ERROR: line 5: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-inspan.pl?q=%3Ciframe%20src=javascript:alert(1)%3B//%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
